todo_list_backend/pyproject.toml
[tool.poetry]
name = "testproject"
version = "0.1.0"
description = ""
authors = ["Yevgeniy Kap <kapustaevg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
pydantic-settings = "^2.5.2"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
uvicorn = "^0.30.6"
django-rq = "^2.10.2"
redis = "^5.0.8"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
